examples/kddcup2021/CITATIONNETWORKV1/r_unimp/requirements.txt
paddle>=2.0.0
pgl>=2.1.2
numpy
tqdm
pyyaml
tensorboardX
easydict
scipy
scikit-learn
